requirements/base.txt
Django==4.2.16
amqp==5.1.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1
django-cors-headers==3.13.0
django-filter==23.3
django-modeltranslation==0.18.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
Pillow==11.0.0
psycopg2-binary==2.9.5
python-dotenv==1.0.1
redis==5.0.1